@@ -40,53 +43,55 @@
         "@radix-ui/react-toggle": "^1.1.0",
         "@radix-ui/react-toggle-group": "^1.1.0",
         "@radix-ui/react-tooltip": "^1.1.2",
-        "@supabase/auth-helpers-nextjs": "^0.8.7",
-        "@supabase/ssr": "latest",
-        "@supabase/supabase-js": "^2.44.2",
         "@types/mdx": "^2.0.13",
-        "autoprefixer": "10.4.17",
-        "axios": "^1.7.2",
+        "mysql2": "^3.11.3",
+        "amazon-cognito-identity-js": "^6.3.12",
         "class-variance-authority": "^0.7.0",
         "clsx": "^2.1.1",
-        "cmdk": "^0.2.1",
+        "cmdk": "^1.0.0",
         "date-fns": "^3.6.0",
-        "dayjs": "^1.11.11",
-        "embla-carousel-react": "8.0.0-rc19",
+        "dayjs": "^1.11.13",
+        "embla-carousel-react": "^8.3.0",
         "feconfig-bhs": "^1.0.7",
-        "geist": "^1.3.0",
-        "isomorphic-dompurify": "^2.13.0",
-        "lucide-react": "^0.312.0",
-        "next": "latest",
-        "next-mdx-remote": "^4.4.1",
-        "next-themes": "^0.2.1",
-        "postcss": "8.4.33",
-        "react": "18.2.0",
+        "input-otp": "^1.2.4",
+        "lucide-react": "^0.411.0",
+        "next": "14.2.5",
+        "next-auth": "5.0.0-beta.20",
+        "next-mdx-remote": "^5.0.0",
+        "next-themes": "^0.3.0",
+        "react": "^18.3.1",
         "react-day-picker": "^8.10.1",
-        "react-dom": "18.2.0",
-        "react-hook-form": "^7.52.1",
-        "react-resizable-panels": "^1.0.10",
-        "rehype": "^13.0.1",
-        "rehype-pretty-code": "^0.13.2",
-        "remark-gfm": "^3.0.1",
-        "sharp": "^0.33.4",
-        "shiki": "^1.10.1",
+        "react-dom": "^18.3.1",
+        "react-hook-form": "^7.53.0",
+        "react-player": "^2.16.0",
+        "react-resizable-panels": "^2.1.4",
+        "recharts": "^2.12.7",
+        "rehype-pretty-code": "^0.14.0",
+        "remark-gfm": "^4.0.0",
+        "screenfull": "^6.0.2",
         "sonner": "^1.5.0",
-        "tailwind-merge": "^2.3.0",
-        "tailwindcss": "3.4.1",
+        "drizzle-orm": "^0.33.0",
+        "tailwind-merge": "^2.5.2",
         "tailwindcss-animate": "^1.0.7",
-        "typescript": "5.3.3",
-        "vaul": "^0.8.9",
-        "zod": "^3.23.8"
+        "unist-util-visit": "^5.0.0",
+        "vaul": "^0.9.9",
+        "zod": "^3.23.8",
+        "zustand": "^4.5.5"
     },
     "devDependencies": {
-        "@tailwindcss/typography": "^0.5.13",
-        "@types/node": "20.11.5",
-        "@types/react": "18.2.48",
-        "@types/react-dom": "18.2.18",
-        "encoding": "^0.1.13",
-        "eslint": "^8.57.0",
-        "eslint-config-next": "14.0.3",
-        "eslint-config-prettier": "^8.10.0",
-        "prettier": "3.2.4"
+        "@tailwindcss/typography": "^0.5.15",
+        "@types/node": "^20.16.10",
+        "@types/react": "^18.3.10",
+        "@types/react-dom": "^18.3.0",
+        "eslint": "^8.57.1",
+        "eslint-config-next": "14.2.5",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-import-resolver-alias": "^1.1.2",
+        "eslint-plugin-prettier": "^5.2.1",
+        "postcss": "^8.4.47",
+        "prettier": "^3.3.3",
+        "tailwindcss": "^3.4.13",
+        "typescript": "^5.6.2",
+        "drizzle-kit": "^0.24.2"
     }
 }
