@@ -1,39 +1,26 @@
-import { NextResponse, type NextRequest } from 'next/server'
-import { createClient } from '@/utils/supabase/middleware'
+import { auth } from '@shared/auth'
 import { headers } from 'next/headers'
+import { NextResponse, type NextRequest } from 'next/server'
 
-const middleware = async (request: NextRequest) => {
+const middleware = async (request: NextRequest, response: NextResponse) => {
+    const session = await auth()
     const headersList = headers()
     const ip = (headersList.get('x-forwarded-for') ?? 'UNKNOWN').split(',')[0]
     const { pathname } = request.nextUrl
 
-    let session
-    let res
-    try {
-        const { supabase, response } = createClient(request)
-        res = response
-        await supabase.from('visitors').insert({ ip, url: pathname })
-        session = await supabase.auth.getSession()
-    } catch (e) {
-        return NextResponse.next({
-            request: {
-                headers: request.headers,
-            },
-        })
-    }
-
-    const authPaths = ['/editor', '/git', '/api/image/upload']
+    const authPaths = ['/write', '/api/image/upload']
     const isAuthPath = authPaths.some((authPath) => pathname.includes(authPath))
-    const isAdmin = session.data.session?.user.email === process.env.ADMIN_EMAIL
+    const isAdmin = session?.user
+
+    
 
     if (isAuthPath && !isAdmin) {
         return NextResponse.redirect(new URL('/login', request.url))
     }
-    return res
 }
 
 export const config = {
-    matcher: ['/((?!_next/static|_next/post|image|favicon.ico).*)'],
+    matcher: ['/((?!_next/static|favicon.ico).*)'],
 }
 
 export default middleware
