@@ -1,10 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    transpilePackages: [
-        'shiki',
-        'next-mdx-remote',
-        'next-mdx-remote/serialize',
-    ],
+    output: 'standalone',
+    transpilePackages: ['shiki', 'next-mdx-remote', 'next-mdx-remote/serialize'],
     images: {
         formats: ['image/avif', 'image/webp'],
         remotePatterns: [
